% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astroCornerplot.R
\name{astroCornerplot}
\alias{astroCornerplot}
\title{Astronomical Cornerplots}
\usage{
astroCornerplot(
  data,
  varlist = names(data),
  contour = TRUE,
  fill = "mediumpurple",
  palette = "Inferno",
  bins = 70,
  contourBins = 5
)
}
\arguments{
\item{data}{a data frame}

\item{varlist}{character, list with all numeric variables to create the cornerplot and histograms with. defaults to all variables in the dataset}

\item{contour}{logical, specify whether to overlay contour plot on 2d histogram}

\item{fill}{fill color for histograms, defaults to mediumpurple}

\item{palette}{color palette for 2d histogram scale, defaults to inferno}

\item{bins}{number of bins for all histograms, defaults to 70}

\item{contourBins}{number of bins for contour plot, defaults to 5}
}
\value{
corner plot with 2d histogram showing intersection of both variables and regular histogram for each variable
}
\description{
Create cornerplots for astronomy data for an unlimited number of paired variables.
}
\examples{
astroCornerplot(starcatalog)
astroCornerplot(starcatalog, contour=FALSE, fill="red4", palette = "Reds", bins=40)
}
