% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astroScatter.R
\name{astroScatter}
\alias{astroScatter}
\title{Astronomical Scatterplots}
\usage{
astroScatter(data, x, y, size = 0.5, alpha = 0.5)
}
\arguments{
\item{data}{a data frame}

\item{x}{a numeric variable}

\item{y}{a numeric variable}
}
\value{
scatter plot with both variables
}
\description{
Create scatterplots for astronomy data, optimized for a large number of points
}
\examples{
astroScatter(starcatalog, vmag, bv_color)
astroScatter(starcatalog, vmag, bv_color, size=0.2, alpha=0.7)
}
