% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cornerplot.R
\name{cornerplot}
\alias{cornerplot}
\title{Cornerplots}
\usage{
cornerplot(
  data,
  varlist = names(data),
  contour = TRUE,
  fill = "mediumpurple",
  palette = "Inferno",
  bins = 30,
  contourBins = 3
)
}
\arguments{
\item{data}{a data frame.}

\item{varlist}{character, list with all numeric variables to create the cornerplot and histograms with. defaults to all variables in the dataset.}

\item{contour}{(optional) logical, specify whether to overlay contour plot on 2d histogram, defaults to TRUE.}

\item{fill}{(optional) fill color for histograms, defaults to mediumpurple.}

\item{palette}{(optional) color palette for 2d histogram scale, defaults to Inferno. options are available through the \code{scale_fill_continuous_sequential} function in the colorspace package.}

\item{bins}{(optional) number of bins for all histograms, defaults to 70.}

\item{contourBins}{(optional) number of bins for contour plot, defaults to 5.}
}
\value{
corner plot with 2d histogram showing intersection of both variables and regular histogram for each variable. Please note that the function can take a while to load, especially if you are running it on more than five variables.
}
\description{
Create cornerplots for numeric data using an unlimited number of paired variables. While you can use an unlimited number of variables, it is not recommended to use more than six numeric variables.
}
\examples{
cornerplot(starcatalog)
cornerplot(starcatalog, contour=FALSE, fill="mediumpurple", palette = "Purples", bins=40)
cornerplot(starcatalog, varlist = c("vmag", "bv_color"), palette = "Blues", fill="dodgerblue3", bins = 50, contourBins = 3)
}
