% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrheatmap.R
\name{corrheatmap}
\alias{corrheatmap}
\title{Correlation Heat Map with or Without Dendrogram}
\usage{
corrheatmap(
  data,
  colors = c("blue", "white", "red"),
  title = "Correlation Heatmap",
  dend = "both"
)
}
\arguments{
\item{data}{a matrix or a data frame}

\item{title}{(optional) a character vector. Default is "Correlation Heatmap"}

\item{dend}{(optional) a string: "both","row","column","none". Default is both.}

\item{color}{(optional) a vector of three colors. Default is blue, white, and red.}
}
\value{
A heatmap graph in a new quartz window.
}
\description{
Calculates Pearson Correlation from a Matrix and plots resulting correlation values on a heatmap with color key and histogram.
}
\examples{
corrheatmap(mtcars, title="Correlations of Car Features", dend="row")
corrheatmap(mtcars, title="Correlations of Car Features", dend="both")
}
