% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_data_test.R
\docType{data}
\name{network_data_test}
\alias{network_data_test}
\title{network_data_test}
\format{
A data frame with 13 rows and 4 variables:
\describe{
\item{\code{from}}{character. DESCRIPTION.}
\item{\code{to}}{character. DESCRIPTION.}
\item{\code{inital_test}}{character. DESCRIPTION.}
\item{\code{target_test}}{character. DESCRIPTION.}
}
}
\usage{
network_data_test
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
