% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplots for Large Amounts of Data}
\usage{
scatterplot(
  data,
  x,
  y,
  size = 0.5,
  alpha = 0.5,
  color = NULL,
  group = NULL,
  palette = "agSunset"
)
}
\arguments{
\item{data}{a data frame}

\item{x}{string, a numeric variable}

\item{y}{string, a numeric variable}

\item{size}{(optional) size for the points on the graph. defaults to 0.5}

\item{alpha}{(optional) transparency for the points on the graph. defaults to 0.5}

\item{color}{(optional) string, color of the points. takes either a variable from the dataset or the name of a color.}

\item{group}{(optional) string, facet variable for the graphs.}

\item{palette}{(optional) string, color palette for points, if color provided is a variable. defaults to agSunset. see \code{scale_color_discrete_sequential} function in the colorspace package for palette options.}
}
\value{
scatter plot with all specified variables
}
\description{
Create scatterplots for astronomy data, optimized for a large number of points
}
\examples{
scatterplot(starcatalog, "vmag", "bv_color")
scatterplot(starcatalog, "vmag", "bv_color", color = "multiple", palette = "Terrain", size = 0.2, alpha = 0.7)
scatterplot(starcatalog, "vmag", "bv_color", color = "mediumpurple", palette = "Terrain", size = 0.2, alpha = 0.7)
scatterplot(starcatalog, "vmag", "bv_color", group = "multiple")
}
