% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciViz-package.R
\docType{package}
\name{SciViz-package}
\alias{SciViz}
\alias{SciViz-package}
\title{SciViz: Creates Publication-Ready Visualizations for Astronomy, Psychology, and Neuroscience Data}
\description{
This package allows scientists to create publication-ready visualizations without having knowledge of how to code graphs in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chenning2011/SciViz}
  \item Report bugs at \url{https://github.com/chenning2011/SciViz/issues}
}

}
\author{
\strong{Maintainer}: Caleb Henning \email{chenning@wesleyan.edu}

Authors:
\itemize{
  \item Angelica Crown \email{acrown@wesleyan.edu}
  \item Larissa Xu \email{lxu01@wesleyan.edu}
}

}
\keyword{internal}
