% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astroScatter(1).R, R/astroScatter.R
\name{astroScatter}
\alias{astroScatter}
\title{Astronomical Scatterplots}
\usage{
astroScatter(data, x, y, size = 0.5, alpha = 0.5)

astroScatter(data, x, y, size = 0.5, alpha = 0.5)
}
\arguments{
\item{data}{a data frame}

\item{x}{a numeric variable}

\item{y}{a numeric variable}
}
\value{
scatter plot with both variables

scatter plot with both variables
}
\description{
Create scatterplots for astronomy data, optimized for a large number of points

Create scatterplots for astronomy data, optimized for a large number of points
}
\examples{
a <- data.frame( x=rnorm(20000, 10, 1.9), y=rnorm(20000, 10, 1.2) )
b <- data.frame( x=rnorm(20000, 14.5, 1.9), y=rnorm(20000, 14.5, 1.9) )
c <- data.frame( x=rnorm(20000, 9.5, 1.9), y=rnorm(20000, 15.5, 1.9) )
data <- rbind(a,b,c)
astroCornerplot(data, x, y)
a <- data.frame( x=rnorm(20000, 10, 1.9), y=rnorm(20000, 10, 1.2) )
b <- data.frame( x=rnorm(20000, 14.5, 1.9), y=rnorm(20000, 14.5, 1.9) )
c <- data.frame( x=rnorm(20000, 9.5, 1.9), y=rnorm(20000, 15.5, 1.9) )
data <- rbind(a,b,c)
astroCornerplot(data, x, y)
}
