% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barData1.R
\docType{data}
\name{barData1}
\alias{barData1}
\title{BarData1}
\format{
A data frame with 10 rows and 4 variables:
\describe{
\item{\code{Group}}{character. DESCRIPTION.}
\item{\code{Time}}{character. DESCRIPTION.}
\item{\code{Depression_Scores}}{double. DESCRIPTION.}
\item{\code{Error_Values}}{double. DESCRIPTION.}
}
}
\usage{
barData1
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
