% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barData2.R
\docType{data}
\name{barData2}
\alias{barData2}
\title{barData2}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{\code{Activity}}{character. DESCRIPTION.}
\item{\code{Session_Type}}{character. DESCRIPTION.}
\item{\code{Mean_Duration}}{double. DESCRIPTION.}
\item{\code{Error_Value}}{double. DESCRIPTION.}
}
}
\usage{
barData2
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
