% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_network_plot.R
\name{interact_network_plot}
\alias{interact_network_plot}
\title{Interactive Network Plot}
\usage{
interact_network_plot(
  data,
  initial_pt = 1,
  target_pt = 2,
  linkColour = "#666",
  nodeColour = "#69b3a2",
  opacity = 0.9,
  title = "Network Graph",
  subtitle = "Exploring Interactions and Relationships"
)
}
\arguments{
\item{data}{a data frame of edge list format. This means each row should specify a link between 2 nodes. There should be two columns; one for the inital point, and a second target point.}

\item{linkColour}{color of edges, MUST be a common color for the whole graph. default is #666}

\item{nodeColour}{color of nodes, MUST be a common color for the whole graph. Default is #69b3a2}

\item{opacity}{opacity of nodes. 0=transparent. 1=no transparency. Default is .9}

\item{title}{default is "Network Graph"}

\item{subtitle}{default is "Exploring Interactions and Relationships"}

\item{to}{column NUMBER of initial point variable. Default is 1}

\item{from}{column NUMBER of target point variable. Default is 2}
}
\value{
an interactive network graph
}
\description{
Create interactive network plot for a data frame of edge list format.
}
\examples{
data_test <- data.frame(
from=c("Neuron 1", "Neuron 1", "Neuron 2", "Neuron 3", "Neuron 5", "Neuron 4", "Neuron 5", "Neuron 7", "Neuron 6", "Neuron 8", "Neuron 9", "Neuron 11", "Neuron 10"),
to=c("Neuron 2", "Neuron 3", "Neuron 1", "Neuron 4", "Neuron 1", "Neuron 6", "Neuron 2", "Neuron 8", "Neuron 2", "Neuron 9", "Neuron 5", "Neuron 7", "Neuron 1"),
inital_test=c("A", "A", "B", "D", "C", "D", "E", "B", "C", "K", "E", "A", "M"),
target_test=c("B", "E", "F", "A", "C", "A", "B", "Z", "A", "C", "A", "B", "K")
)
interact_network_plot(data_test)
interact_network_plot(data_test, initial_pt = 3, target_pt = 4)

}
