% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineData1.R
\docType{data}
\name{lineData1}
\alias{lineData1}
\title{lineData1}
\format{
A data frame with 15 rows and 4 variables:
\describe{
\item{\code{Time}}{integer. DESCRIPTION.}
\item{\code{Value}}{double. DESCRIPTION.}
\item{\code{Group}}{character. DESCRIPTION.}
\item{\code{Error}}{double. DESCRIPTION.}
}
}
\usage{
lineData1
}
\description{
This dataset contains time-series values across three groups, with associated error values for each time point, enabling the visualization of trends and variability over time.
}
\keyword{datasets}
