% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_data_test.R
\docType{data}
\name{network_data_test}
\alias{network_data_test}
\title{network_data_test}
\format{
A data frame with 13 rows and 4 variables:
\describe{
\item{\code{from}}{character. DESCRIPTION.}
\item{\code{to}}{character. DESCRIPTION.}
\item{\code{inital_test}}{character. DESCRIPTION.}
\item{\code{target_test}}{character. DESCRIPTION.}
}
}
\usage{
network_data_test
}
\description{
network_data_test includes 4 columns, therefore 2 sets of network data in one data frame. There are 13 rows of random data. Columns 1 and 2 are of the form "Neuron 1" -> "Neuron 2". Columns 3 and 4 are of the form "A" -> "B".
}
\keyword{datasets}
